requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ML
tensorflow==2.16.1
scikit-learn==1.4.2
numpy==1.26.4
joblib==1.4.0
opencv-python-headless==4.9.0.80
Pillow==10.3.0

# Supabase
supabase==2.4.3

# Scheduler (autonomous retraining)
APScheduler==3.10.4

# Env vars
python-dotenv==1.0.1

# UI
streamlit==1.35.0
matplotlib==3.9.0
pandas==2.2.2
requests==2.32.2

# Load testing (install separately with: pip install locust)
# locust==2.28.0
